+++ .travis.yml
@@ -9,3 +9,4 @@ script:
 - set -o pipefail && cd Example && bundle install && bundle exec pod install && cd ../ && xcodebuild build -workspace Example/Example.xcworkspace -scheme Example -sdk iphonesimulator ONLY_ACTIVE_ARCH=NO | xcpretty
 - pod lib lint --quick
 - carthage build --platform iOS --no-skip-current
+- swift build -Xswiftc "-sdk" -Xswiftc "`xcrun --sdk iphonesimulator --show-sdk-path`" -Xswiftc "-target" -Xswiftc "x86_64-apple-ios14.0-simulator"
